支持模态窗模整个应用进程意见修改2

+++ interfaces/innerkits/wm/wm_common.h
@@ -1204,7 +1204,7 @@ struct SystemWindowOptions {
     int32_t windowType = -1;
 };
 
-enum class ModalityType : uint32_t {
+enum class ModalityType : uint8_t {
     WINDOW_MODALITY,
     APPLICATION_MODALITY,
 };
